engines rework; code cleanup part1;

+++ src/net/sf/l2j/gameserver/handler/skillhandlers/Disablers.java
@@ -104,7 +104,7 @@ public class Disablers implements ISkillHandler {
                                 break;
                             case AGGDAMAGE:
                                 if (target instanceof Attackable)
-                                    target.getAI().notifyEvent(AiEventType.AGGRESSION, activeChar, Integer.valueOf((int) (150.0D * skill.getPower() / (target.getLevel() + 7))));
+                                    target.getAI().notifyEvent(AiEventType.AGGRESSION, activeChar, (int) (150.0D * skill.getPower() / (target.getLevel() + 7)));
                                 skill.getEffects(activeChar, target, new Env(shld, ss, sps, bsps));
                                 break;
                             case AGGREDUCE:
